requirements.txt
Django==5.2.10
python-decouple==3.8
dj-database-url==2.2.0
whitenoise==6.7.0
gunicorn==22.0.0
psycopg[binary]==3.2.1

# LLM API Libraries (Optional - install based on which service you choose)
# Groq API (Free, fastest)
groq==0.9.0
httpx==0.27.2

# Image + ranking dependencies
requests==2.32.3
python-dotenv==1.0.1
sentence-transformers==3.0.1
scikit-learn==1.5.2


